removing websocket stuff

+++ batch/sort_file.py
@@ -7,39 +7,6 @@ import json
 import dask.dataframe as dd
 import pandas as pd
 import re
-import websockets
-import asyncio
-
-connected_clients = set()
-
-
-async def handle_client(websocket, path):
-    connected_clients.add(websocket)
-    try:
-        async for message in websocket:
-            print(f"Received message from client: {message}")
-
-    finally:
-        # Remove client from the set when disconnected
-        connected_clients.remove(websocket)
-
-
-async def send_message_to_all_clients(message):
-    # Send a message to all connected clients
-    if connected_clients:  # Check if there are any connected clients
-        await asyncio.gather(
-            *[client.send(message) for client in connected_clients]
-        )
-
-
-async def server():
-    async with websockets.serve(handle_client, "0.0.0.0", 5000) as ws_server:
-        try:
-            await asyncio.Future()
-        except asyncio.CancelledError:
-            await asyncio.gather(*(ws.close() for ws in connected_clients))
-            raise
-        return ws_server
 
 
 def download_file_from_s3(s3_path):
@@ -101,8 +68,6 @@ def csv_to_jsonl(csv_file_path, jsonl_file_path):
 @click.option('--schema', '-a', type=str, required=True)
 @click.option('--already_sorted', '-o', type=bool, required=True)
 def main(s3_path, columns_to_sort, schema, already_sorted):
-    loop = asyncio.get_event_loop()
-    server_task = loop.create_task(server())
     sorted_file = None
     schema_info = json.loads(schema)
     columns_to_sort = columns_to_sort.split(',')
@@ -110,29 +75,19 @@ def main(s3_path, columns_to_sort, schema, already_sorted):
     local_file = download_file_from_s3(s3_path)
     if not already_sorted:
         print("Sorting file")
-        loop.run_until_complete(send_message_to_all_clients("SORTING FILE"))
-        sorted_file = sort_file(local_file, columns_to_sort, schema_info)
 
     json_file = local_file[:-3] + 'json'
     print("Converting to json")
-    loop.run_until_complete(send_message_to_all_clients("CONVERTING TO JSON"))
     csv_to_jsonl(local_file if already_sorted else sorted_file, json_file)
 
     print("Uploading json to s3")
-    loop.run_until_complete(send_message_to_all_clients("UPLOADING JSON"))
     upload_file_to_s3(json_file, s3_path)
     print("finished")
-    loop.run_until_complete(send_message_to_all_clients("READY TO INDEX"))
 
-    server_task.cancel()
-    try:
-        loop.run_until_complete(server_task)
-    except asyncio.CancelledError:
-        pass
-
-        # Clean up local files
+    # Clean up local files
     os.remove(local_file)
-    os.remove(sorted_file)
+    if sorted_file:
+        os.remove(sorted_file)
     os.remove(json_file)
